Move `contact/index.ts` to `contact/post.ts` and adjust routes for 404
Fixes https://github.com/zeit/now/issues/3160.

The `methods` array for a route is not exclusive. Meaning that if the
`src` matches but the method does not match in the `methods` array, then
the routes processing continues on to the next route definition, rather
than causing a 404 for that route.

Now specifies implicit routes when the endpoint is named `index.*`, so
that `GET /api/contact` routes to `/api/contact/index.ts` implicitly,
and that is the route that gets matched after the `POST` one is skipped.
To avoid this, the `index.ts` file is renamed to `post.ts` so that the
implicit route is not created.

Additionally, `GET /api/contact/post.ts` also does work, so we have to
add an additional route after the `POST` one to cause that request to
404. You may also define your own custom 404 page at `/404` if you
desire, otherwise the default Now 404 page will be used.